+++ tests/test_logical_links.py
@@ -110,9 +110,9 @@ class TestLogicalLinks(TestCase):
             discharge_sum = 0.0
 
             for i_conn, (_pipe, orientation) in connected_pipes.items():
-                discharge_sum += results[f"{node}.GasConn[{i_conn+1}].Q"] * orientation
+                discharge_sum += results[f"{node}.GasConn[{i_conn + 1}].Q"] * orientation
                 np.testing.assert_allclose(
-                    results[f"{node}.GasConn[{i_conn+1}].H"], results[f"{node}.H"], atol=1.0e-6
+                    results[f"{node}.GasConn[{i_conn + 1}].H"], results[f"{node}.H"], atol=1.0e-6
                 )
             np.testing.assert_allclose(discharge_sum, 0.0, atol=1.0e-12)
 
@@ -149,9 +149,9 @@ class TestLogicalLinks(TestCase):
             discharge_sum = 0.0
 
             for i_conn, (_pipe, orientation) in connected_pipes.items():
-                discharge_sum += results[f"{node}.HeatConn[{i_conn+1}].Q"] * orientation
+                discharge_sum += results[f"{node}.HeatConn[{i_conn + 1}].Q"] * orientation
                 np.testing.assert_allclose(
-                    results[f"{node}.HeatConn[{i_conn+1}].H"], results[f"{node}.H"], atol=1.0e-6
+                    results[f"{node}.HeatConn[{i_conn + 1}].H"], results[f"{node}.H"], atol=1.0e-6
                 )
             np.testing.assert_allclose(discharge_sum, 0.0, atol=1.0e-12)